pyproject.toml
[project]
name = "ocr_trim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "httpx>=0.28.1",
    "lmstudio>=1.5.0",
    "matplotlib>=3.10.7",
    "modelscope>=1.31.0",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "structured-logprobs>=0.1.5",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
